+++ nautobot/apps/ui.py
@@ -1,10 +1,14 @@
 """Utilities for apps to integrate with and extend the existing Nautobot UI."""
 
-from nautobot.core.apps import (
+from nautobot.core.choices import ButtonColorChoices
+from nautobot.core.ui.base import PermissionsMixin
+from nautobot.core.ui.homepage import (
     HomePageBase,
     HomePageGroup,
     HomePageItem,
     HomePagePanel,
+)
+from nautobot.core.ui.nav import (
     NavContext,
     NavGrouping,
     NavItem,
@@ -15,9 +19,7 @@ from nautobot.core.apps import (
     NavMenuImportButton,
     NavMenuItem,
     NavMenuTab,
-    PermissionsMixin,
 )
-from nautobot.core.choices import ButtonColorChoices
 from nautobot.extras.choices import BannerClassChoices
 from nautobot.extras.plugins import Banner, TemplateExtension
